Get rid of a separate ExecutionTraceModule (replaced by ExecutionTrace)
