+++ radix-engine/src/engine/mod.rs
@@ -2,7 +2,7 @@ mod actor;
 mod auth_module;
 mod call_frame;
 mod errors;
-mod execution_trace_module;
+mod execution_trace;
 mod kernel;
 mod native_interpreter;
 mod node;
@@ -18,7 +18,7 @@ pub use actor::*;
 pub use auth_module::*;
 pub use call_frame::CallFrame;
 pub use errors::*;
-pub use execution_trace_module::*;
+pub use execution_trace::*;
 pub use kernel::*;
 pub use native_interpreter::NativeInterpreter;
 pub use node::*;
